+++ src/steps/was/types.ts
@@ -1,0 +1,6 @@
+export type Description = {
+  recommendation?: string;
+  reference?: string;
+  description?: string;
+  impact?: string;
+};