+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+### Added
+
+- Added `recommendation`, `reference`, `description`, and `impact` fields to
+  `qualys_web_app_finding`
+
+### Changed
+
+- Changed severities to be uppercased words (e.g. `critical` -> `Critical`).
+
 ## [5.13.0] - 2022-11-03
 
 ### Added