+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [5.14.0] - 2023-04-07
+
 ### Changed
 
 - New properties added to resources: